requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning
torch>=1.9.0
tensorflow>=2.6.0

# Explainability
shap>=0.40.0
lime>=0.2.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Interface
gradio>=3.0.0
streamlit>=1.0.0

# Data Processing
pandas-profiling>=3.0.0
feature-engine>=1.2.0

# Model Persistence
joblib>=1.0.0
pickle-mixin>=1.0.0

# Configuration
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipywidgets>=7.6.0
notebook>=6.4.0

# Development and Testing
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
pre-commit>=2.15.0

# Utilities
tqdm>=4.62.0
click>=8.0.0
pathlib2>=2.3.6